Test case was testing for NoSuchProcess exceptions when accessing kill() and pid - pid is never going to throw an exception since it's the original parameter the Process object is created with, and kill() currently does not throw a NoSuchProcess exception (at least not on Windows and OS X).
Windows platform now checks for a pid in the process list before fetching get_process_info() and raises a NoSuchProcess exception. If OpenProcess fails we throw an exception using the last Windows error (i.e. whatever is returned from GetLastError). 

Possibly we need a better way to check if a process is valid than checking in the PID list, I *think* it's possible that a PID is still listed in the process list but by the time we do our system calls to read the name/commandline etc. it can return invalid results because the process is no longer active/valid. Need to look into this on all platforms, essentially how to handle a "zombie" process.



git-svn-id: http://psutil.googlecode.com/svn/trunk@134 d8a29897-7e4b-0410-a74a-8bb4c295346f