Additional changes for Issue #50: http://code.google.com/p/psutil/issues/detail?id=50
Attached is another version for the same psutil/arch/osx/process_info.c file with modifications in the GetBSDProcessList and getcmdargs functions.

The GetBSDProcessList functions is simpler and uses an increased buffer size to improve the chances for 
success on the second sysctl call.

Function getcmdargs uses the same double sysctl call method as GetBSDProcessList and only get the 
KERN_ARGMAX value as a last resort.  In addition, getcmdargs avoids fragmenting memory for smaller 
argument lists by using the space on the stack.




git-svn-id: http://psutil.googlecode.com/svn/trunk@395 d8a29897-7e4b-0410-a74a-8bb4c295346f